+++ docs/assets/navigation.js
@@ -1,1 +1,1 @@
-window.navigationData = "eJyllM1ugzAQhF+l2jNq2vRP5ZZDuFJVaS8RBxeMQBgbrU1TVOXdK9wkNcFgR716Z77dYY2336Dol4IQopanqhRcQgBpUbIMKYdwe6pTRAigIaqAEPKjeEERrwtVMwigKnkG4eP9PjiZchT1SnY83RQoduSDURtjrHIgnTQfkKhsblHNWJDKlqkIRW2z/lW9EC8o6lJaI4xEM8CW75A0+vPF1g0NBL6gNbNPNhL5AvVGnEStciLncnpEdKXzDObMNBUnMVCbrqFXK1YSSad+vLg5/paHPqprqFwcjof05c3z0+3D0mjwqq/Sufn31M+rV2MH6NIFlIjPcCLuR1qbD5FJWZ8/RVOEqTF8J4gruz+u7H5z4e8Ey/5hmtq2JLlxNT+P6kV/PqTfLQNIGZESQshogzQlimZg9Oo9Z9sbAi37c1Nb3lvf9BW3gc36f9iTk49Efl2SffIDHaFfNg=="
+window.navigationData = "eJyllUtPwzAQhP8K2nNEoLxEbhVqrkFQTlUPJtkqURw7WjuUCvW/o5o+nNZxDVw9M9/6MZJnX6DxU0MCT5IQIsjLiheEApLZXkIiiKBluoQEFp3IdSWFipHostQNhwjqShSQ3N+uo31I1q6MrE8icyv0uhL5wCYUW+CB+MGoYu8cVbxZ7zNvRhHknCkFCRTYEuZMYwHWmE5sUm9iSax1MW39t+wFyWZaklxuUK4r6Bk8F0ioOq5Tko2LclA9iM6cIHO+3k4LiE+4cp7E1gMw5tA+jjF4+zFW/oLsdGdLjPifqgzST0x/KY0Jnm1O3xVUn2eSTeV+wRPT2Wc0832N2hpCQf5uWaZQ4JmW2S5v1aarFi/GvGIK1UDjstaQD8P0qkUVb5f79NHV48P13cga8GLu/jj8sxqWPeqjDXC00EtJhYeTijDSxP4nbMrk+KcYIgxtI3QHWe3OZ7U7P1/PvwFt72Uy"
